refactor: domain 기반 패키지 구조 변경 및 기존 코드 정리
