+++ src/main/java/com/leets/blog/global/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.leets.blog.config;
+package com.leets.blog.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
