+++ src/main/java/com/leets/blog/post/dto/PostResponse.java
@@ -1,0 +1,17 @@
+package com.leets.blog.post.dto;
+
+import com.leets.blog.post.domain.Post;
+import lombok.Getter;
+
+@Getter
+public class PostResponse {
+    private Long id;
+    private String title;
+    private String content;
+
+    public PostResponse(Post post) {
+        this.id = post.getId();
+        this.title = post.getTitle();
+        this.content = post.getContent();
+    }
+}
