+++ src/main/java/com/leets/blog/post/repository/PostRepository.java
@@ -1,0 +1,9 @@
+package com.leets.blog.post.repository;
+
+import com.leets.blog.post.domain.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
