+++ src/main/java/com/leets/blog/user/domain/UserRole.java
@@ -1,13 +1,14 @@
-package com.leets.blog.domain.post;
+package com.leets.blog.user.domain;
+
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum PostStatus {
-    DRAFT("임시 저장"),
-    PUBLISHED("발행 완료");
+public enum UserRole {
+    USER("일반 사용자"),
+    ADMIN("관리자");
 
     private final String description;
 }
